+++ docs/about/changelog.rst
@@ -1,6 +1,33 @@
 Changelog
 ==========
 
+Version *Next*
+---------------
+
+**Released**: TBD
+
+**Features**
+
+- Improve plagiarism document rendering `(#1335)
+  <https://github.com/CodeGra-de/CodeGra.de/pull/1335>`__. Matching blocks of
+  code can now be rendered side by side, the amount of context lines before and
+  after each match is configurable, and it is possible to export to a docx
+  file.
+
+**Fixes**
+
+- Miscellaneous fixes:
+  `(#1322) <https://github.com/CodeGra-de/CodeGra.de/pull/1322>`__,
+  `(#1323) <https://github.com/CodeGra-de/CodeGra.de/pull/1323>`__,
+  `(#1324) <https://github.com/CodeGra-de/CodeGra.de/pull/1324>`__,
+  `(#1325) <https://github.com/CodeGra-de/CodeGra.de/pull/1325>`__,
+  `(#1326) <https://github.com/CodeGra-de/CodeGra.de/pull/1326>`__,
+  `(#1327) <https://github.com/CodeGra-de/CodeGra.de/pull/1327>`__,
+  `(#1328) <https://github.com/CodeGra-de/CodeGra.de/pull/1328>`__,
+  `(#1331) <https://github.com/CodeGra-de/CodeGra.de/pull/1331>`__,
+  `(#1332) <https://github.com/CodeGra-de/CodeGra.de/pull/1332>`__,
+  `(#1333) <https://github.com/CodeGra-de/CodeGra.de/pull/1333>`__.
+
 Version LowVoltage
 -------------------